Share validation logic between Claims and EDIFACT order validators
Extract the common validate() skeleton into AbstractExportParametersValidator,
parameterized by getExpectedIntegrationType() and a validateSpecific() hook.
Also rename two stale BursarFeeFines test names in EDIFACT validator test.
